+++ src/Yasumi/data/translations/mothersDay.php
@@ -17,11 +17,14 @@ declare(strict_types = 1);
 
 // Translations for Mother's Day
 return [
+    'da' => 'mors dag',
     'de' => 'Muttertag',
     'el' => 'Ημέρα της μητέρας',
     'en' => 'Mother’s Day',
     'fr' => 'Fête des mères',
     'it' => 'Festa della mamma',
+    'ka' => 'დედის დღე',
+    'lv' => 'Mātes dienu',
     'nl' => 'Moederdag',
     'pt' => 'Dia da Mãe',
     'ro' => 'Ziua Mamei',