+++ tests/Latvia/PentecostTest.php
@@ -1,0 +1,83 @@
+<?php
+
+declare(strict_types = 1);
+
+/**
+ * This file is part of the 'Yasumi' package.
+ *
+ * The easy PHP Library for calculating holidays.
+ *
+ * Copyright (c) 2015 - 2025 AzuyaLabs
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Sacha Telgenhof <me at sachatelgenhof dot com>
+ */
+
+namespace Yasumi\tests\Latvia;
+
+use Yasumi\Holiday;
+use Yasumi\tests\HolidayTestCase;
+
+/**
+ * Class containing tests for Pentecost in Latvia.
+ */
+class PentecostTest extends LatviaBaseTestCase implements HolidayTestCase
+{
+    /**
+     * The name of the holiday to be tested.
+     */
+    public const HOLIDAY = 'pentecost';
+
+    /**
+     * @return array<array> list of test dates for the holiday defined in this test
+     *
+     * @throws \Exception
+     */
+    public function holidayDataProvider(): array
+    {
+        return $this->generateRandomPentecostDates(self::TIMEZONE);
+    }
+
+    /**
+     * Test defined holiday in the test.
+     *
+     * @dataProvider holidayDataProvider
+     *
+     * @param int    $year     the year for which the holiday defined in this test needs to be tested
+     * @param string $expected the expected date
+     *
+     * @throws \Exception
+     */
+    public function testHoliday(int $year, string $expected): void
+    {
+        $this->assertHoliday(
+            self::REGION,
+            self::HOLIDAY,
+            $year,
+            new \DateTime($expected, new \DateTimeZone(self::TIMEZONE))
+        );
+    }
+
+    /**
+     * Tests the translated name of the holiday defined in this test.
+     */
+    public function testTranslation(): void
+    {
+        $this->assertTranslatedHolidayName(
+            self::REGION,
+            self::HOLIDAY,
+            $this->generateRandomYear(),
+            [self::LOCALE => 'Vasarsvētkus']
+        );
+    }
+
+    /**
+     * Tests type of the holiday defined in this test.
+     */
+    public function testHolidayType(): void
+    {
+        $this->assertHolidayType(self::REGION, self::HOLIDAY, $this->generateRandomYear(), Holiday::TYPE_OFFICIAL);
+    }
+}